added ability to add patients and patient notes for Doctors and Pharmacists

+++ MediHelp/src/main/java/com/se452/group5/MediHelp/entity/Patient.java
@@ -19,7 +19,7 @@ public class Patient implements Serializable {
         
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long PATIENTID;
+    private int PATIENTID;
 
     private Long User_ID;
